Refactor NanoGPT provider hooks
