+++ package.json
@@ -32,6 +32,11 @@
     "runtime/usage.ts",
     "catalog/models-json-snapshot.ts",
     "catalog/build-provider.ts",
+    "provider/auth.ts",
+    "provider/catalog-hooks.ts",
+    "provider/tool-schema-hooks.ts",
+    "provider/error-hooks.ts",
+    "provider/stream-hooks.ts",
     "provider-catalog.ts",
     "provider-discovery.ts",
     "nanogpt-errors.ts",
